feat(node): Add kurtosis e2e testing skeleton (#1792)
## Description
This PR adds the necessary infrastructure to automatically deploy and
run e2e tests for the node with Kurtosis.

Next PR will add that to CI to make sure we don't break the node tests
when adding new features.

Note that I had to port the mise config from the ethereum optimism.