fix(node): Zero `safe` + `finalized` hashes while EL syncs (#1801)
## Overview

Sets the `safe` and `finalized` hashes to `B256::ZERO` while the EL
syncs, and also only finalizes the unsafe ref if the client is
undergoing its _initial_ sync.

Also moves the priority of the `sync_complete_rx.recv()` future in the
biased select within the derivation actor. Since the futures are polled
in-order, that future always resolves once the channel has been closed -
it must be placed last.