fix(node/p2p): fix node id serialization and connectedness display (#1824)
## Description

Small PR that fixes the `NodeId` serialization inside `opp2p_self`.
Using the default `Display` impl strips the middle of the address for
readability. We want to be able to return the full address in RPC

Also removes the special display for `Connectedness` which was causing
the RPC responses for `opp2p_self` returned by `kona-node` to differ
from ones from the `op-node` (only return a `u8`)