feat(node/p2p): streaming engine state through websockets (#1833)
## Description

This PR adds a couple of websocket endpoints to stream the sync state of
the `kona-node`. These endpoints can be useful inside our e2e testing
suite to ensure that the node is correctly syncing.

## Follow-ups

Maybe we should provide a configuration flag to make those RPC endpoints
optional. We may not want to be able to subscribe to the current node
state. Here is a ticket to track that #1842

Close #1825