chore(node/net): integrate the network driver inside the network actor (#2376)
## Description

This PR refactors the network driver and moves it inside the network
actor. This reduces the levels of indirection and allow the network
actor to comply more with the new builder refactor of the node services
by avoiding to define channels inside the driver.

Changes:

- Removes the `net` module inside `kona-p2p`. Moves the relevant files
to the `kona-node-services` crate.
- Removes the `broadcast` module inside the network actor. This can now
be directly handled using a mpsc channel inside the network actor to
implement the notion of queue. Outer interactions with the network actor
happen through the `NetworkActor` context (see for instance
https://github.com/op-rs/kona/compare/main...theo/send-payloads-to-network#diff-b4e04fd02334c1ae45e956ad21312e81efacba5a7c1d477b176711dbae458c32R87-R126
or
https://github.com/op-rs/kona/compare/main...theo/send-payloads-to-network#diff-00c31e9f9de711af31ebbd3b81f610ea12ea100a6a8d0db6a2ef86f52f298849L80-R75)
- Divides the network actor module with the following pattern to
propagate the lower level abstractions (resp
`Discv5Driver`/`Discv5Handler`/`GossipDriver`):
```
NodeBuilder ->build NodeDriver ->(async) start NodeHandler
```

Close #2365