feat(node/sequencer): implement remote signer skeleton (#2572)
## Description

This PR spikes the remote signer for kona node's sequencer. In
particular it:
- Introduces a new `BlockSigner` enum which encapsulates the behavior of
a `LocalSigner` and a `RemoteSigner`
- Propagates the changes of the `BlockSigner` enum along the stack.
Still use the `LocalSigner` everywhere
- Adds the core of the remote signer code. In particular, it ensures we
implement the `op-signer` interface, we handle CA and client
certificates and support hot reloads for certificates through a file
watcher (like the op-node does)

Follow-up: add CLI options to activate remote signer from the command
line.